factor: remove `array_sharded_ext` module
